create cached property out of nargs
